Restructure nav, home, about pages
