+++ frontend/src/pages/Home.jsx
@@ -1,48 +1,20 @@
 // src/pages/Home.jsx
 import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import "../App.css";
+import NavigationBar from "../components/nav/NavigationBar";
+import "../styles/home/home.css";
 
 function Home() {
   const navigate = useNavigate();
-  //For Dropdown
-  const [selected, setSelected] = useState("");
-  const handleChange = (e) => {
-    const url = e.target.value;
-    if (url) {
-      window.location.href = url; // Redirects to the selected link
-    }
-  };
 
   return (
     <div className="App">
       <header className="App-header">
-        <nav>
-          <h1>PADDY PLATFORM</h1>
-          <ul>
-            <li>
-              <Link to="/">Home</Link>
-            </li>
-            <li>
-              <Link to="/about">About</Link>
-            </li>
-            <li>
-              <select value={selected} onChange={handleChange}>
-                <option value="">Features</option>
-                <option value="/features/weather">Weather</option>
-                <option value="/features/chat">Chat</option>
-                <option value="/features/chatbot">Chatbot</option>
-              </select>
-            </li>
-            <li><Link to="/gigs">Gigs</Link></li>
-            <li><Link to="/shop">Shop</Link></li>
-            <li><Link to="/blog">Blog</Link></li>
-            <li><Link to="/login">Login</Link></li>
-          </ul>
-        </nav>
+        <NavigationBar />
 
         <div className="hero">
-          <Link to="/login">become a Seller/Buyer</Link>
+          <Link to="/login">become a Seller/Buyer</Link> <br />
+          <Link to="/about">Why Us?</Link>
         </div>
       </header>
     </div>
